[Refactor] 로또 개수 확인 메서드 controller로 이동
